cloud_functions/oauth/mailchimp-callback/requirements.txt
functions-framework==3.*
requests==2.*
google-cloud-secret-manager==2.*
google-cloud-firestore==2.*
